refactor type + findSegmentIndex
